add dashboard functionality backend for instructor

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection = require("../models/SubSection");
 // CREATE a new section
 exports.createSection = async (req, res) => {
 	try {
@@ -54,15 +55,26 @@ exports.createSection = async (req, res) => {
 // UPDATE a section
 exports.updateSection = async (req, res) => {
 	try {
-		const { sectionName, sectionId } = req.body;
+		const { sectionName, sectionId,courseId } = req.body;
 		const section = await Section.findByIdAndUpdate(
 			sectionId,
 			{ sectionName },
 			{ new: true }
 		);
+
+		const course = await Course.findById(courseId)
+		.populate({
+			path:"courseContent",
+			populate:{
+				path:"subSection",
+			},
+		})
+		.exec();
+
 		res.status(200).json({
 			success: true,
 			message: section,
+			data:course,
 		});
 	} catch (error) {
 		console.error("Error updating section:", error);
@@ -76,11 +88,40 @@ exports.updateSection = async (req, res) => {
 // DELETE a section
 exports.deleteSection = async (req, res) => {
 	try {
-		const { sectionId } = req.body;
+
+		const { sectionId, courseId }  = req.body;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
+
+		//delete sub section
+		await SubSection.deleteMany({_id: {$in: section.subSection}});
+
 		await Section.findByIdAndDelete(sectionId);
+
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
+
 		res.status(200).json({
-			success: true,
-			message: "Section deleted",
+			success:true,
+			message:"Section deleted",
+			data:course
 		});
 	} catch (error) {
 		console.error("Error deleting section:", error);
@@ -89,4 +130,4 @@ exports.deleteSection = async (req, res) => {
 			message: "Internal server error",
 		});
 	}
-};
+};
